links now showed as results if filtered
